Moved common methods in class PacmanGraphics from captureGraphicsDisplay and graphicsDisplay into ABC absPacmanGraphics in module topGraphics.
